tree editor

+++ js/input.js
@@ -1,9 +1,10 @@
 let data = [{
-        id: 1,
-        text_1: "Root",
-        father: null
-    }
-];
+    id: 1,
+    text_1: "Root",
+    father: null,
+    selected: true,
+    subselected: false, 
+}];
 
 let currentMaxID = 1;
 
@@ -18,20 +19,64 @@ let myTree = Treeviz.create({
     mainAxisNodeSpacing: 2,
     isHorizontal: false,
     renderNode: function (node) {
-        return result = "<div onclick='onNodeClick(this)' data-id=" + node.data.id + " class='box' style='cursor:pointer;height:" + node.settings.nodeHeight + "px; width:" + node.settings.nodeWidth + "px;display:flex;flex-direction:column;justify-content:center;align-items:center;background-color:" +
-            node.data.color +
-            ";border-radius:5px;'><div><strong>" +
-            node.data.text_1 +
-            "</strong></div><div>is</div><div><i>" +
-            node.data.text_2 +
-            "</i></div></div>";
+        return result = getNodeElement(node);
     },
     linkWidth: (nodeData) => 5,
     linkShape: "curve",
     linkColor: (nodeData) => "#B0BEC5",
 });
 
-myTree.refresh(data);
+function getNodeElement(node) {
+    const styleString = `
+    cursor: pointer;
+    height: ${node.settings.nodeHeight}px; 
+    width: ${node.settings.nodeWidth}px; 
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    border-radius: 5px;
+  `;
+
+    let classList = 'card';
+    if (node.data.selected == true) {
+        classList += ' border-success bg-success text-light';
+    }
+    if (node.data.subselected == true) {
+        classList += ' border-danger bg-danger text-light';
+    }
+
+    const htmlString = `
+    <div 
+      onclick='onNodeClick(this)' 
+      data-id="${node.data.id}" 
+      class="${classList}"
+      style="${styleString}"
+    >
+      <div>
+        <strong>${node.data.text_1}</strong>
+      </div>
+    </div>
+  `;
+
+    return htmlString;
+}
+
+function selectById(id) {
+    for (let i = 0; i < data.length; i++) {
+        if (data[i].id == id) {
+            data[i].selected = true;
+        } else {
+            data[i].selected = false;
+        }
+        if (data[i].father == id) {
+            data[i].subselected = true;
+        } else {
+            data[i].subselected = false;
+        }
+    }
+    myTree.refresh(data);
+}
 
 function getElementById(id) {
     return data.find(item => item.id === id);
@@ -42,86 +87,105 @@ function onNodeClick(node) {
     let container = document.querySelector('#level-0');
     container.innerHTML = "";
     container.appendChild(drawLevel(id));
+    selectById(id);
 }
 
 function updateEditorById(id) {
+    if (id == null) return; 
+
     let container = document.querySelector('#level-0');
     container.innerHTML = "";
     container.appendChild(drawLevel(id));
+    selectById(id);
 }
 
 function getBorderClass(level) {
-    let border_classes = ["border-primary", "border-success", "border-danger", "border-warning", "border-info"];
-    return border_classes[level % border_classes.length];
+    let borderClasses = ["border-primary", "border-success", "border-danger", "border-warning", "border-info"];
+    return borderClasses[level % borderClasses.length];
 }
 
 function drawLevel(id) {
     const element = data.find(item => item.id == id);
     if (!element) return '';
 
-    let new_div = createNode(element, 1);
+    let newDiv = createNode(element, 1);
 
     const children = data.filter(item => item.father == id);
     for (const child of children) {
-        let child_div = createNode(child, 2);
-        new_div.lastChild.appendChild(child_div);
+        let childDiv = createNode(child, 2);
+        newDiv.lastChild.appendChild(childDiv);
     }
 
-    return new_div;
+    return newDiv;
 }
 
 function createNode(node, borderLevel) {
-    let new_div = document.createElement('div');
-    new_div.classList.add('node', 'card');
-    new_div.classList.add(getBorderClass(borderLevel));
+    let newDiv = document.createElement('div');
+    newDiv.classList.add('node', 'card');
+    newDiv.classList.add(getBorderClass(borderLevel));
 
-    let card_header = document.createElement('div');
-    card_header.classList.add("card-header");
-    card_header.textContent = node.text_1;
-    new_div.appendChild(card_header);
+    let cardHeader = document.createElement('div');
+    cardHeader.classList.add("card-header")
+    cardHeader.setAttribute('role', 'button');
+    cardHeader.textContent = node.text_1;
+    newDiv.appendChild(cardHeader);
 
-    let new_node_input = document.createElement('div');
-    new_node_input.classList.add('node-input', 'mt-2');
+    let newNodeInput = document.createElement('div');
+    newNodeInput.classList.add('node-input', 'mt-2');
 
-    let new_input = document.createElement('input');
-    new_input.placeholder = 'Название этого уровня';
-    new_input.classList.add('form-control');
-    new_node_input.appendChild(new_input);
+    let newInput = document.createElement('input');
+    newInput.placeholder = 'Название этого уровня';
+    newInput.classList.add('form-control');
+    newInput.addEventListener("input", (event) => {
+        node.text_1 = event.target.value;
+        cardHeader.textContent = event.target.value;
+        myTree.refresh(data);
+    });
+    newNodeInput.appendChild(newInput);
 
-    let new_btn_group = document.createElement('div');
-    new_btn_group.classList.add('buttons');
+    let newBtnGroup = document.createElement('div');
+    newBtnGroup.classList.add('buttons');
 
-    let new_button = document.createElement('button');
-    new_button.setAttribute('onclick', 'addLevel(' + node.id + ')');
-    new_button.classList.add('btn', 'btn-primary');
-    new_button.innerHTML = '<i class="fa-solid fa-plus"></i>';
-    new_btn_group.appendChild(new_button);
+    let newButton = document.createElement('button');
+    newButton.setAttribute('onclick', 'addLevel(' + node.id + ')');
+    newButton.classList.add('btn', 'btn-primary');
+    newButton.innerHTML = '<i class="fa-solid fa-plus"></i>';
+    newBtnGroup.appendChild(newButton);
 
-    let delete_button = document.createElement('button');
-    delete_button.setAttribute('onclick', 'deleteLevel(' + node.id + ')');
-    delete_button.classList.add('btn', 'btn-danger');
-    delete_button.innerHTML = '<i class="fa-solid fa-trash"></i>';
-    new_btn_group.appendChild(delete_button);
+    let deleteButton = document.createElement('button');
+    deleteButton.setAttribute('onclick', 'deleteLevel(' + node.id + ')');
+    deleteButton.classList.add('btn', 'btn-danger');
+    deleteButton.innerHTML = '<i class="fa-solid fa-trash"></i>';
+    newBtnGroup.appendChild(deleteButton);
 
-    new_node_input.appendChild(new_btn_group);
+    newNodeInput.appendChild(newBtnGroup);
 
-    new_div.appendChild(new_node_input);
+    newDiv.appendChild(newNodeInput);
 
-    let new_childs_div = document.createElement('div');
-    new_childs_div.classList.add('node-childs');
-    new_div.appendChild(new_childs_div);
+    let newChildrenDiv = document.createElement('div');
+    newChildrenDiv.classList.add('node-childs');
+    newDiv.appendChild(newChildrenDiv);
 
-    new_div.addEventListener('mouseover', function () {
-        let computed_style = window.getComputedStyle(new_div);
-        let border_color = computed_style.getPropertyValue('border-top-color');
-        new_div.style.boxShadow = "0 0 5px 3px " + border_color;
+    newDiv.addEventListener('mouseover', function () {
+        let computedStyle = window.getComputedStyle(newDiv);
+        let borderColor = computedStyle.getPropertyValue('border-top-color');
+        newDiv.style.boxShadow = "0 0 5px 3px " + borderColor;
     });
 
-    new_div.addEventListener('mouseout', function () {
-        new_div.style.boxShadow = "none";
+    newDiv.addEventListener('mouseout', function () {
+        newDiv.style.boxShadow = "none";
     });
 
-    return new_div;
+    cardHeader.addEventListener('click', (event) => {
+        if (borderLevel == 1) {
+            updateEditorById(node.father);
+        }
+        if (borderLevel == 2) {
+            updateEditorById(node.id);
+        }
+    });
+
+    return newDiv;
 }
 
 function addLevel(parentId) {
@@ -130,8 +194,8 @@ function addLevel(parentId) {
         text_1: "No name",
         father: parentId
     });
-    updateEditorById(parentId);
     myTree.refresh(data);
+    updateEditorById(parentId);
 }
 
 function deleteById(data, id) {
@@ -146,12 +210,17 @@ function deleteById(data, id) {
         }
         index = data.findIndex(item => item.id == id);
     }
-    
+
 }
 
 function deleteLevel(id) {
+    if (id == 1) {
+        alert("Невозможно удалить корневой элемент");
+        return
+    }
     let parentId = getElementById(id).father;
     deleteById(data, id);
     updateEditorById(parentId);
-    myTree.refresh(data);
 }
+
+updateEditorById(1);
